proyecto-core/requirements.txt
# Base
sqlalchemy>=2.0.0
pymysql>=1.0.2
python-dotenv>=0.20.0
pytest>=7.0.0
pytest-html>=4.0.0

# Procesamiento de documentos
python-docx>=0.8.11
PyPDF2>=2.10.0
pandas>=1.5.0
openpyxl>=3.0.10
python-magic>=0.4.27
langdetect>=1.0.9

# API y Web
fastapi>=0.92.0
uvicorn>=0.20.0

# IA
openai>=0.27.0
deepseek>=0.0.4

# Herramientas de desarrollo
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.0.0
pytest-cov>=4.0.0